Fix some clang-tidy warnings

+++ src/actions.c
@@ -331,7 +331,7 @@ edit_actions(char *app)
 		free(paths);
 	}
 
-	path_n = (size_t)get_path_env(1);
+	path_n = get_path_env(1);
 	get_path_programs();
 
 	print_reload_msg(_("File modified. Actions reloaded\n"));
